feat(rest): Extract proxy tests
