+++ connectors/http/http-base/src/test/java/io/camunda/connector/http/base/client/apache/CustomApacheHttpClientTest.java
@@ -22,10 +22,8 @@ import static com.github.tomakehurst.wiremock.client.WireMock.any;
 import static com.github.tomakehurst.wiremock.client.WireMock.containing;
 import static com.github.tomakehurst.wiremock.client.WireMock.created;
 import static com.github.tomakehurst.wiremock.client.WireMock.delete;
-import static com.github.tomakehurst.wiremock.client.WireMock.deleteRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
-import static com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.noContent;
 import static com.github.tomakehurst.wiremock.client.WireMock.notFound;
 import static com.github.tomakehurst.wiremock.client.WireMock.ok;
@@ -65,12 +63,9 @@ import io.camunda.document.store.InMemoryDocumentStore;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.commons.text.StringEscapeUtils;
-import org.apache.hc.client5.http.impl.classic.HttpClients;
 import org.apache.hc.core5.http.ContentType;
 import org.apache.hc.core5.http.HttpHeaders;
 import org.apache.hc.core5.http.HttpStatus;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -78,13 +73,9 @@ import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
-import org.testcontainers.containers.BindMode;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.utility.DockerImageName;
 import wiremock.com.fasterxml.jackson.databind.node.JsonNodeFactory;
-import wiremock.com.fasterxml.jackson.databind.node.POJONode;
 
-@WireMockTest(httpPort = 28090)
+@WireMockTest
 public class CustomApacheHttpClientTest {
 
   private final CustomApacheHttpClient customApacheHttpClient = CustomApacheHttpClient.getDefault();
@@ -130,105 +121,7 @@ public class CustomApacheHttpClientTest {
                       .withName("document")
                       .withBody(equalTo("The content of this file"))
                       .build()));
-    }
-  }
-
-  @Nested
-  class ProxyTests {
-    private static CustomApacheHttpClient proxiedApacheHttpClient;
-    private static GenericContainer<?> proxyContainer;
-
-    @BeforeAll
-    public static void setUp() {
-      proxyContainer =
-          new GenericContainer<>(DockerImageName.parse("sameersbn/squid:3.5.27-2"))
-              .withExposedPorts(3128)
-              .withClasspathResourceMapping(
-                  "squid.conf", "/etc/squid/squid.conf", BindMode.READ_ONLY)
-              .waitingFor(org.testcontainers.containers.wait.strategy.Wait.forListeningPort());
-      proxyContainer.start();
-
-      // Set up the HttpClient to use the proxy
-      String proxyHost = proxyContainer.getHost();
-      Integer proxyPort = proxyContainer.getMappedPort(3128);
-      System.setProperty("http.proxyHost", proxyHost);
-      System.setProperty("http.proxyPort", proxyPort.toString());
-      System.setProperty("http.nonProxyHosts", "");
-      System.setProperty("https.proxyHost", proxyHost);
-      System.setProperty("https.proxyPort", proxyPort.toString());
-      System.setProperty("https.nonProxyHosts", "");
-      proxiedApacheHttpClient = CustomApacheHttpClient.create(HttpClients.custom());
-    }
-
-    @AfterAll
-    public static void tearDown() {
-      proxyContainer.stop();
-      System.setProperty("http.proxyHost", "");
-      System.setProperty("http.proxyPort", "");
-      System.setProperty("http.nonProxyHosts", "");
-      System.setProperty("https.proxyHost", "");
-      System.setProperty("https.proxyPort", "");
-      System.setProperty("https.nonProxyHosts", "");
-    }
-
-    @Test
-    public void shouldReturn200_whenGetAndProxySet(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(get("/path").willReturn(ok().withBody("Hello, world!")));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setUrl("http://host.docker.internal:28090/path");
-      HttpCommonResult result = proxiedApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      assertThat(result.body()).isEqualTo("Hello, world!");
-      assertThat(result.headers().get("Via")).asString().contains("squid");
-      verify(getRequestedFor(urlEqualTo("/path")));
-    }
-
-    @Test
-    public void shouldReturn200_whenPostAndProxySet(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(
-          post("/path").willReturn(created().withJsonBody(new POJONode(Map.of("key1", "value1")))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setUrl("http://host.docker.internal:28090/path");
-      HttpCommonResult result = proxiedApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-      assertThat(result.body()).isEqualTo(Map.of("key1", "value1"));
-      assertThat(result.headers().get("Via")).asString().contains("squid");
-      verify(postRequestedFor(urlEqualTo("/path")));
-    }
-
-    @Test
-    public void shouldReturn200_whenPutAndProxySet(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(put("/path").willReturn(ok().withJsonBody(new POJONode(Map.of("key1", "value1")))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.PUT);
-      request.setUrl("http://host.docker.internal:28090/path");
-      HttpCommonResult result = proxiedApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      assertThat(result.body()).isEqualTo(Map.of("key1", "value1"));
-      assertThat(result.headers().get("Via")).asString().contains("squid");
-      verify(putRequestedFor(urlEqualTo("/path")));
-    }
-
-    @Test
-    public void shouldReturn200_whenDeleteAndProxySet(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(delete("/path").willReturn(noContent()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.DELETE);
-      request.setUrl("http://host.docker.internal:28090/path");
-      HttpCommonResult result = proxiedApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(204);
-      assertThat(result.headers().get("Via")).asString().contains("squid");
-      verify(deleteRequestedFor(urlEqualTo("/path")));
+      store.deleteDocument(ref);
     }
   }
 
@@ -239,6 +132,7 @@ public class CustomApacheHttpClientTest {
     @EnumSource(HttpMethod.class)
     public void shouldReturn200_whenSpaceInPathAndQueryParameters(
         HttpMethod method, WireMockRuntimeInfo wmRuntimeInfo) {
+
       stubFor(any(urlEqualTo("/path%20with%20spaces?andQuery=S%C3%A3o%20Paulo")).willReturn(ok()));
 
       HttpCommonRequest request = new HttpCommonRequest();
@@ -287,702 +181,715 @@ public class CustomApacheHttpClientTest {
     }
   }
 
-  @Nested
-  class GetTests {
-
-    @Test
-    public void shouldReturn200_whenNullHeaders(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(get("/path").willReturn(ok()));
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      var headers = new HashMap<String, String>();
-      headers.put("Content-Type", null);
-      request.setHeaders(headers);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-    }
-
-    @Test
-    public void shouldReturn200_whenNoTimeouts(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(get("/path").willReturn(ok()));
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setConnectionTimeoutInSeconds(null);
-      request.setReadTimeoutInSeconds(null);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-    }
-
-    @Test
-    public void shouldReturn200WithoutBody_whenEmptyGet(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(get("/path").willReturn(ok()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenGetWithBody(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(get("/path").willReturn(ok("Hello, world!")));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      assertThat(result.body()).isEqualTo("Hello, world!");
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenGetWithBodyJSON(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(
-          get("/path")
-              .willReturn(
-                  ok().withJsonBody(
-                          JsonNodeFactory.instance
-                              .objectNode()
-                              .put("name", "John")
-                              .put("age", 30)
-                              .putNull("message"))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setHeaders(Map.of("Accept", "application/json"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      JSONAssert.assertEquals(
-          "{\"name\":\"John\",\"age\":30,\"message\":null}",
-          objectMapper.writeValueAsString(result.body()),
-          JSONCompareMode.STRICT);
-    }
-
-    @ParameterizedTest
-    @EnumSource(HttpMethod.class)
-    public void shouldReturn200WithBody_whenGetWithBodyXML(
-        HttpMethod method, WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      stubFor(
-          any(urlEqualTo("/path?format=xml"))
-              .willReturn(
-                  ok().withBody(
-                          "<note>\n"
-                              + "  <to>Tove</to>\n"
-                              + "  <from>Jani</from>\n"
-                              + "  <heading>Reminder</heading>\n"
-                              + "  <body>Don't forget me this weekend!</body>\n"
-                              + "</note>")));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(method);
-      request.setQueryParameters(Map.of("format", "xml"));
-      request.setHeaders(Map.of("Accept", "application/xml"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      assertThat(result.body())
-          .isEqualTo(
-              "<note>\n"
-                  + "  <to>Tove</to>\n"
-                  + "  <from>Jani</from>\n"
-                  + "  <heading>Reminder</heading>\n"
-                  + "  <body>Don't forget me this weekend!</body>\n"
-                  + "</note>");
-    }
-
-    @Test
-    public void shouldReturn500_whenGetWithInvalidBody(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(get("/path").willReturn(serverError().withStatusMessage("Invalid JSON")));
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      ConnectorException e =
-          assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
-      assertThat(e.getErrorCode()).isEqualTo("500");
-      assertThat(e.getMessage()).contains("Invalid JSON");
-    }
-
-    @Test
-    public void shouldReturn404_whenGetWithNonExistingPath(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(get("/path").willReturn(notFound().withBody("Not Found: /path")));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      ConnectorException e =
-          assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
-      assertThat(e.getErrorCode()).isEqualTo("404");
-      assertThat(e.getMessage()).contains("Not Found");
-      assertThat(((Map) e.getErrorVariables().get("response")).get("body"))
-          .isEqualTo("Not Found: /path");
-    }
-
-    @Test
-    public void shouldReturn408_whenGetWithTimeout(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(get("/path").willReturn(ok().withFixedDelay(2000)));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      request.setReadTimeoutInSeconds(1);
-      ConnectorException e =
-          assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
-      assertThat(e.getErrorCode()).isEqualTo(String.valueOf(HttpStatus.SC_REQUEST_TIMEOUT));
-      assertThat(e.getMessage())
-          .contains("An error occurred while executing the request, or the connection was aborted");
-    }
-  }
-
-  @Nested
-  class PostTests {
-
-    @Test
-    public void shouldReturn201WithoutBody_whenEmptyPost(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(post("/path").willReturn(created()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-    }
-
-    @Test
-    public void shouldReturn201WithBody_whenPostBody(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(post("/path").willReturn(created()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setHeaders(Map.of("header", "headerValue"));
-      request.setBody(Map.of("key1", "value1"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-
-      verify(
-          postRequestedFor(urlEqualTo("/path"))
-              .withHeader("Content-Type", equalTo("application/json"))
-              .withHeader("header", equalTo("headerValue"))
-              .withRequestBody(equalTo(StringEscapeUtils.unescapeJson("{\"key1\":\"value1\"}"))));
-    }
-
-    @Test
-    public void shouldReturn201WithBody_whenPostBodyURLEncoded(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(post("/path").willReturn(created()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setHeaders(
-          Map.of(HttpHeaders.CONTENT_TYPE, ContentType.APPLICATION_FORM_URLENCODED.getMimeType()));
-      request.setBody(Map.of("key1", "value1", "key2", "value2"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-
-      verify(
-          postRequestedFor(urlEqualTo("/path"))
-              .withHeader("Content-Type", equalTo("application/x-www-form-urlencoded"))
-              .withRequestBody(
-                  and(containing("key1=value1"), containing("&"), containing("key2=value2"))));
-    }
-
-    @Test
-    public void shouldReturn201WithBody_whenPostBodyMultiPart(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(post("/path").withMultipartRequestBody(aMultipart()).willReturn(created()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setHeaders(
-          Map.of(HttpHeaders.CONTENT_TYPE, ContentType.MULTIPART_FORM_DATA.getMimeType()));
-      request.setBody(Map.of("key1", "value1", "key2", "value2"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-
-      verify(
-          postRequestedFor(urlEqualTo("/path"))
-              .withHeader(
-                  "Content-Type", and(containing("multipart/form-data"), containing("boundary=")))
-              .withRequestBodyPart(
-                  new MultipartValuePatternBuilder()
-                      .withName("key1")
-                      .withBody(equalTo("value1"))
-                      .build())
-              .withRequestBodyPart(
-                  new MultipartValuePatternBuilder()
-                      .withName("key2")
-                      .withBody(equalTo("value2"))
-                      .build()));
-    }
-
-    @Test
-    public void shouldReturn201WithBody_whenPostBodyMultiPartWithBoundaryProvided(
-        WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(post("/path").withMultipartRequestBody(aMultipart()).willReturn(created()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setHeaders(
-          Map.of(
-              HttpHeaders.CONTENT_TYPE,
-              "multipart/form-data; charset=ISO-8859-1; boundary=g7wNbtOKHnEq4vnSoWdDYS88OICfGHzBA68DqmJS"));
-      request.setBody(Map.of("key1", "value1", "key2", "value2"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-
-      verify(
-          postRequestedFor(urlEqualTo("/path"))
-              .withHeader(
-                  "Content-Type",
-                  and(
-                      containing("multipart/form-data"),
-                      containing("boundary=g7wNbtOKHnEq4vnSoWdDYS88OICfGHzBA68DqmJS")))
-              .withRequestBodyPart(
-                  new MultipartValuePatternBuilder()
-                      .withName("key1")
-                      .withBody(equalTo("value1"))
-                      .build())
-              .withRequestBodyPart(
-                  new MultipartValuePatternBuilder()
-                      .withName("key2")
-                      .withBody(equalTo("value2"))
-                      .build()));
-    }
-
-    @Test
-    public void shouldReturn201WithBody_whenPostBodyTextPlainWithStringBody(
-        WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      stubFor(post("/path").willReturn(created()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE, ContentType.TEXT_PLAIN.getMimeType()));
-      request.setBody("Hello, world!");
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-
-      verify(
-          postRequestedFor(urlEqualTo("/path"))
-              .withHeader("Content-Type", equalTo("text/plain"))
-              .withRequestBody(equalTo("Hello, world!")));
-    }
-
-    @Test
-    public void shouldReturn201WithBody_whenPostBodyTextPlainWithIntegerBody(
-        WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      stubFor(post("/path").willReturn(created()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE, ContentType.TEXT_PLAIN.getMimeType()));
-      request.setBody(123);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-
-      verify(
-          postRequestedFor(urlEqualTo("/path"))
-              .withHeader("Content-Type", equalTo("text/plain"))
-              .withRequestBody(equalTo("123")));
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenPostBodyTextPlainWithBooleanBody(
-        WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      stubFor(post("/path").willReturn(created()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.POST);
-      request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE, ContentType.TEXT_PLAIN.getMimeType()));
-      request.setBody(true);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(201);
-
-      verify(
-          postRequestedFor(urlEqualTo("/path"))
-              .withHeader("Content-Type", equalTo("text/plain"))
-              .withRequestBody(equalTo("true")));
-    }
-  }
-
-  @Nested
-  class DeleteTests {
-    @Test
-    public void shouldReturn204WithoutBody_whenDelete(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(delete("/path/id").willReturn(noContent()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.DELETE);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path/id");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.body()).isNull();
-      assertThat(result.status()).isEqualTo(204);
-    }
-  }
-
-  @Nested
-  class PutTests {
-    @Test
-    public void shouldReturn200WithoutBody_whenEmptyPut(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(put("/path").willReturn(ok()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.PUT);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenPutBody(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(put("/path").willReturn(ok()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.PUT);
-      request.setHeaders(Map.of("header", "headerValue"));
-      request.setBody(Map.of("key1", "value1"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-
-      verify(
-          putRequestedFor(urlEqualTo("/path"))
-              .withHeader("Content-Type", equalTo("application/json"))
-              .withHeader("header", equalTo("headerValue"))
-              .withRequestBody(equalTo(StringEscapeUtils.unescapeJson("{\"key1\":\"value1\"}"))));
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenPutBodyURLEncoded(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(put("/path").willReturn(ok()));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.PUT);
-      request.setHeaders(
-          Map.of(HttpHeaders.CONTENT_TYPE, ContentType.APPLICATION_FORM_URLENCODED.getMimeType()));
-      request.setBody(Map.of("key1", "value1", "key2", "value2"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-
-      verify(
-          putRequestedFor(urlEqualTo("/path"))
-              .withHeader("Content-Type", equalTo("application/x-www-form-urlencoded"))
-              .withRequestBody(
-                  and(containing("key1=value1"), containing("&"), containing("key2=value2"))));
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenPutBodyTextPlainWithStringBody(
-        WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      stubFor(put("/path").willReturn(ok().withBody("Hello, world updated!")));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.PUT);
-      request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE, ContentType.TEXT_PLAIN.getMimeType()));
-      request.setBody("Hello, world!");
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      assertThat(result.body()).isEqualTo("Hello, world updated!");
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenPutBodyTextPlainWithIntegerBody(
-        WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      stubFor(put("/path").willReturn(ok().withBody("123")));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.PUT);
-      request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE, ContentType.TEXT_PLAIN.getMimeType()));
-      request.setBody(123);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      assertThat(result.body()).isEqualTo("123");
-    }
-  }
-
-  @Nested
-  class AuthenticationTests {
-
-    @Test
-    public void shouldReturn200WithBody_whenGetWithBasicAuth(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(
-          get("/path")
-              .withBasicAuth("user", "password")
-              .willReturn(
-                  ok().withJsonBody(
-                          JsonNodeFactory.instance
-                              .objectNode()
-                              .put("name", "John")
-                              .put("age", 30)
-                              .putNull("message"))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setHeaders(Map.of("Accept", "application/json"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      request.setAuthentication(new BasicAuthentication("user", "password"));
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      JSONAssert.assertEquals(
-          "{\"name\":\"John\",\"age\":30,\"message\":null}",
-          objectMapper.writeValueAsString(result.body()),
-          JSONCompareMode.STRICT);
-    }
-
-    @Test
-    public void shouldReturn401_whenGetWithWrongBasicAuth(WireMockRuntimeInfo wmRuntimeInfo) {
-      stubFor(
-          get("/path")
-              .withBasicAuth("user", "password")
-              .willReturn(unauthorized().withBody("Unauthorized")));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      request.setAuthentication(new BasicAuthentication("user", "password"));
-      ConnectorException e =
-          assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
-      assertThat(e.getErrorCode()).isEqualTo("401");
-      assertThat(e.getMessage()).contains("Unauthorized");
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenGetWithBearerAuth(WireMockRuntimeInfo wmRuntimeInfo)
-        throws Exception {
-      stubFor(
-          get("/path")
-              .withHeader("Authorization", equalTo("Bearer token"))
-              .willReturn(
-                  ok().withJsonBody(
-                          JsonNodeFactory.instance
-                              .objectNode()
-                              .put("name", "John")
-                              .put("age", 30)
-                              .putNull("message"))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setAuthentication(new BearerAuthentication("token"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      JSONAssert.assertEquals(
-          "{\"name\":\"John\",\"age\":30,\"message\":null}",
-          objectMapper.writeValueAsString(result.body()),
-          JSONCompareMode.STRICT);
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenGetWithApiKeyAuthInHeaders(
-        WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      stubFor(
-          get("/path")
-              .withHeader("theName", equalTo("theValue"))
-              .willReturn(
-                  ok().withJsonBody(
-                          JsonNodeFactory.instance
-                              .objectNode()
-                              .put("name", "John")
-                              .put("age", 30)
-                              .putNull("message"))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setAuthentication(
-          new ApiKeyAuthentication(ApiKeyLocation.HEADERS, "theName", "theValue"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      JSONAssert.assertEquals(
-          "{\"name\":\"John\",\"age\":30,\"message\":null}",
-          objectMapper.writeValueAsString(result.body()),
-          JSONCompareMode.STRICT);
-    }
-
-    @Test
-    public void shouldReturn200WithBody_whenGetWithApiKeyAuthInQueryParams(
-        WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      stubFor(
-          get(urlPathEqualTo("/path"))
-              .withQueryParam("theName", equalTo("theValue"))
-              .willReturn(
-                  ok().withJsonBody(
-                          JsonNodeFactory.instance
-                              .objectNode()
-                              .put("name", "John")
-                              .put("age", 30)
-                              .putNull("message"))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setAuthentication(
-          new ApiKeyAuthentication(ApiKeyLocation.QUERY, "theName", "theValue"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      JSONAssert.assertEquals(
-          "{\"name\":\"John\",\"age\":30,\"message\":null}",
-          objectMapper.writeValueAsString(result.body()),
-          JSONCompareMode.STRICT);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {OAuthConstants.BASIC_AUTH_HEADER, OAuthConstants.CREDENTIALS_BODY})
-    public void shouldReturn200WithBody_whenGetWithOAuthAndCredentialsInBody(
-        String credentialsLocation, WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      createAuthServer(credentialsLocation);
-      stubFor(
-          get("/path")
-              .withHeader("Authorization", equalTo("Bearer token"))
-              .willReturn(
-                  ok().withJsonBody(
-                          JsonNodeFactory.instance
-                              .objectNode()
-                              .put("name", "John")
-                              .put("age", 30)
-                              .putNull("message"))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setAuthentication(
-          new OAuthAuthentication(
-              wmRuntimeInfo.getHttpBaseUrl() + "/oauth",
-              "clientId",
-              "clientSecret",
-              "theAudience",
-              credentialsLocation,
-              "read:resource"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      HttpCommonResult result = customApacheHttpClient.execute(request);
-      assertThat(result).isNotNull();
-      assertThat(result.status()).isEqualTo(200);
-      JSONAssert.assertEquals(
-          "{\"name\":\"John\",\"age\":30,\"message\":null}",
-          objectMapper.writeValueAsString(result.body()),
-          JSONCompareMode.STRICT);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {OAuthConstants.BASIC_AUTH_HEADER, OAuthConstants.CREDENTIALS_BODY})
-    public void shouldReturn401_whenGetWithOAuthReturns401(
-        String credentialsLocation, WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
-      createFailingAuthServer(credentialsLocation);
-      stubFor(
-          get("/path")
-              .withHeader("Authorization", equalTo("Bearer token"))
-              .willReturn(
-                  ok().withJsonBody(
-                          JsonNodeFactory.instance
-                              .objectNode()
-                              .put("name", "John")
-                              .put("age", 30)
-                              .putNull("message"))));
-
-      HttpCommonRequest request = new HttpCommonRequest();
-      request.setMethod(HttpMethod.GET);
-      request.setAuthentication(
-          new OAuthAuthentication(
-              wmRuntimeInfo.getHttpBaseUrl() + "/oauth",
-              "clientId",
-              "clientSecret",
-              "theAudience",
-              credentialsLocation,
-              "read:resource"));
-      request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
-      var e = assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
-      assertThat(e).isNotNull();
-      assertThat(e.getErrorCode()).isEqualTo("401");
-      assertThat(e.getMessage()).contains("Unauthorized");
-    }
-
-    private void createAuthServer(String credentialsLocation) {
-      var request =
-          post("/oauth")
-              .withHeader(
-                  HttpHeaders.CONTENT_TYPE,
-                  equalTo(ContentType.APPLICATION_FORM_URLENCODED.getMimeType()))
-              .withFormParam("grant_type", equalTo("client_credentials"))
-              .withFormParam("audience", equalTo("theAudience"))
-              .withFormParam("scope", equalTo("read:resource"));
-      if (OAuthConstants.CREDENTIALS_BODY.equals(credentialsLocation)) {
-        request
-            .withFormParam("client_id", equalTo("clientId"))
-            .withFormParam("client_secret", equalTo("clientSecret"));
-      } else {
-        request.withBasicAuth("clientId", "clientSecret");
+    @Nested
+    class GetTests {
+
+      @Test
+      public void shouldReturn200_whenNullHeaders(WireMockRuntimeInfo wmRuntimeInfo) {
+        stubFor(get("/path").willReturn(ok()));
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        var headers = new HashMap<String, String>();
+        headers.put("Content-Type", null);
+        request.setHeaders(headers);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
       }
-      stubFor(
-          request.willReturn(
-              ok().withJsonBody(
-                      JsonNodeFactory.instance
-                          .objectNode()
-                          .put("access_token", "token")
-                          .put("token_type", "Bearer")
-                          .put("expires_in", 3600))));
+
+      @Test
+      public void shouldReturn200_whenNoTimeouts(WireMockRuntimeInfo wmRuntimeInfo) {
+        stubFor(get("/path").willReturn(ok()));
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setConnectionTimeoutInSeconds(null);
+        request.setReadTimeoutInSeconds(null);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+      }
+
+      @Test
+      public void shouldReturn200WithoutBody_whenEmptyGet(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(get("/path").willReturn(ok()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenGetWithBody(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(get("/path").willReturn(ok("Hello, world!")));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        assertThat(result.body()).isEqualTo("Hello, world!");
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenGetWithBodyJSON(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(
+            get("/path")
+                .willReturn(
+                    ok().withJsonBody(
+                            JsonNodeFactory.instance
+                                .objectNode()
+                                .put("name", "John")
+                                .put("age", 30)
+                                .putNull("message"))));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setHeaders(Map.of("Accept", "application/json"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        JSONAssert.assertEquals(
+            "{\"name\":\"John\",\"age\":30,\"message\":null}",
+            objectMapper.writeValueAsString(result.body()),
+            JSONCompareMode.STRICT);
+      }
+
+      @ParameterizedTest
+      @EnumSource(HttpMethod.class)
+      public void shouldReturn200WithBody_whenGetWithBodyXML(
+          HttpMethod method, WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        stubFor(
+            any(urlEqualTo("/path?format=xml"))
+                .willReturn(
+                    ok().withBody(
+                            "<note>\n"
+                                + "  <to>Tove</to>\n"
+                                + "  <from>Jani</from>\n"
+                                + "  <heading>Reminder</heading>\n"
+                                + "  <body>Don't forget me this weekend!</body>\n"
+                                + "</note>")));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(method);
+        request.setQueryParameters(Map.of("format", "xml"));
+        request.setHeaders(Map.of("Accept", "application/xml"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        assertThat(result.body())
+            .isEqualTo(
+                "<note>\n"
+                    + "  <to>Tove</to>\n"
+                    + "  <from>Jani</from>\n"
+                    + "  <heading>Reminder</heading>\n"
+                    + "  <body>Don't forget me this weekend!</body>\n"
+                    + "</note>");
+      }
+
+      @Test
+      public void shouldReturn500_whenGetWithInvalidBody(WireMockRuntimeInfo wmRuntimeInfo) {
+        stubFor(get("/path").willReturn(serverError().withStatusMessage("Invalid JSON")));
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        ConnectorException e =
+            assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
+        assertThat(e.getErrorCode()).isEqualTo("500");
+        assertThat(e.getMessage()).contains("Invalid JSON");
+      }
+
+      @Test
+      public void shouldReturn404_whenGetWithNonExistingPath(WireMockRuntimeInfo wmRuntimeInfo) {
+        stubFor(get("/path").willReturn(notFound().withBody("Not Found: /path")));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        ConnectorException e =
+            assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
+        assertThat(e.getErrorCode()).isEqualTo("404");
+        assertThat(e.getMessage()).contains("Not Found");
+        assertThat(((Map) e.getErrorVariables().get("response")).get("body"))
+            .isEqualTo("Not Found: /path");
+      }
+
+      @Test
+      public void shouldReturn408_whenGetWithTimeout(WireMockRuntimeInfo wmRuntimeInfo) {
+        stubFor(get("/path").willReturn(ok().withFixedDelay(2000)));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        request.setReadTimeoutInSeconds(1);
+        ConnectorException e =
+            assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
+        assertThat(e.getErrorCode()).isEqualTo(String.valueOf(HttpStatus.SC_REQUEST_TIMEOUT));
+        assertThat(e.getMessage())
+            .contains("An error occurred while executing the request, or the connection was aborted");
+      }
     }
 
-    private void createFailingAuthServer(String credentialsLocation) {
-      var request =
-          post("/oauth")
-              .withHeader(
-                  HttpHeaders.CONTENT_TYPE,
-                  equalTo(ContentType.APPLICATION_FORM_URLENCODED.getMimeType()))
-              .withFormParam("grant_type", equalTo("client_credentials"))
-              .withFormParam("audience", equalTo("theAudience"))
-              .withFormParam("scope", equalTo("read:resource"));
-      if (OAuthConstants.CREDENTIALS_BODY.equals(credentialsLocation)) {
-        request
-            .withFormParam("client_id", equalTo("clientId"))
-            .withFormParam("client_secret", equalTo("clientSecret"));
-      } else {
-        request.withBasicAuth("clientId", "clientSecret");
+    @Nested
+    class PostTests {
+
+      @Test
+      public void shouldReturn201WithoutBody_whenEmptyPost(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(post("/path").willReturn(created()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.POST);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(201);
       }
-      stubFor(request.willReturn(unauthorized().withBody("Unauthorized")));
+
+      @Test
+      public void shouldReturn201WithBody_whenPostBody(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(post("/path").willReturn(created()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.POST);
+        request.setHeaders(Map.of("header", "headerValue"));
+        request.setBody(Map.of("key1", "value1"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(201);
+
+        verify(
+            postRequestedFor(urlEqualTo("/path"))
+                .withHeader("Content-Type", equalTo("application/json"))
+                .withHeader("header", equalTo("headerValue"))
+
+   .withRequestBody(equalTo(StringEscapeUtils.unescapeJson("{\"key1\":\"value1\"}"))));
+      }
+
+      @Test
+      public void shouldReturn201WithBody_whenPostBodyURLEncoded(WireMockRuntimeInfo
+   wmRuntimeInfo)
+          throws Exception {
+        stubFor(post("/path").willReturn(created()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.POST);
+        request.setHeaders(
+            Map.of(HttpHeaders.CONTENT_TYPE,
+   ContentType.APPLICATION_FORM_URLENCODED.getMimeType()));
+        request.setBody(Map.of("key1", "value1", "key2", "value2"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(201);
+
+        verify(
+            postRequestedFor(urlEqualTo("/path"))
+                .withHeader("Content-Type", equalTo("application/x-www-form-urlencoded"))
+                .withRequestBody(
+                    and(containing("key1=value1"), containing("&"), containing("key2=value2"))));
+      }
+
+      @Test
+      public void shouldReturn201WithBody_whenPostBodyMultiPart(WireMockRuntimeInfo wmRuntimeInfo)
+   {
+        stubFor(post("/path").withMultipartRequestBody(aMultipart()).willReturn(created()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.POST);
+        request.setHeaders(
+            Map.of(HttpHeaders.CONTENT_TYPE, ContentType.MULTIPART_FORM_DATA.getMimeType()));
+        request.setBody(Map.of("key1", "value1", "key2", "value2"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(201);
+
+        verify(
+            postRequestedFor(urlEqualTo("/path"))
+                .withHeader(
+                    "Content-Type", and(containing("multipart/form-data"),
+   containing("boundary=")))
+                .withRequestBodyPart(
+                    new MultipartValuePatternBuilder()
+                        .withName("key1")
+                        .withBody(equalTo("value1"))
+                        .build())
+                .withRequestBodyPart(
+                    new MultipartValuePatternBuilder()
+                        .withName("key2")
+                        .withBody(equalTo("value2"))
+                        .build()));
+      }
+
+      @Test
+      public void shouldReturn201WithBody_whenPostBodyMultiPartWithBoundaryProvided(
+          WireMockRuntimeInfo wmRuntimeInfo) {
+        stubFor(post("/path").withMultipartRequestBody(aMultipart()).willReturn(created()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.POST);
+        request.setHeaders(
+            Map.of(
+                HttpHeaders.CONTENT_TYPE,
+                "multipart/form-data; charset=ISO-8859-1; boundary=g7wNbtOKHnEq4vnSoWdDYS88OICfGHzBA68DqmJS"));
+        request.setBody(Map.of("key1", "value1", "key2", "value2"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(201);
+
+        verify(
+            postRequestedFor(urlEqualTo("/path"))
+                .withHeader(
+                    "Content-Type",
+                    and(
+                        containing("multipart/form-data"),
+                        containing("boundary=g7wNbtOKHnEq4vnSoWdDYS88OICfGHzBA68DqmJS")))
+                .withRequestBodyPart(
+                    new MultipartValuePatternBuilder()
+                        .withName("key1")
+                        .withBody(equalTo("value1"))
+                        .build())
+                .withRequestBodyPart(
+                    new MultipartValuePatternBuilder()
+                        .withName("key2")
+                        .withBody(equalTo("value2"))
+                        .build()));
+      }
+
+      @Test
+      public void shouldReturn201WithBody_whenPostBodyTextPlainWithStringBody(
+          WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        stubFor(post("/path").willReturn(created()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.POST);
+        request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE,
+   ContentType.TEXT_PLAIN.getMimeType()));
+        request.setBody("Hello, world!");
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(201);
+
+        verify(
+            postRequestedFor(urlEqualTo("/path"))
+                .withHeader("Content-Type", equalTo("text/plain"))
+                .withRequestBody(equalTo("Hello, world!")));
+      }
+
+      @Test
+      public void shouldReturn201WithBody_whenPostBodyTextPlainWithIntegerBody(
+          WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        stubFor(post("/path").willReturn(created()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.POST);
+        request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE,
+   ContentType.TEXT_PLAIN.getMimeType()));
+        request.setBody(123);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(201);
+
+        verify(
+            postRequestedFor(urlEqualTo("/path"))
+                .withHeader("Content-Type", equalTo("text/plain"))
+                .withRequestBody(equalTo("123")));
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenPostBodyTextPlainWithBooleanBody(
+          WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        stubFor(post("/path").willReturn(created()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.POST);
+        request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE,
+   ContentType.TEXT_PLAIN.getMimeType()));
+        request.setBody(true);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(201);
+
+        verify(
+            postRequestedFor(urlEqualTo("/path"))
+                .withHeader("Content-Type", equalTo("text/plain"))
+                .withRequestBody(equalTo("true")));
+      }
     }
-  }
+
+    @Nested
+    class DeleteTests {
+      @Test
+      public void shouldReturn204WithoutBody_whenDelete(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(delete("/path/id").willReturn(noContent()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.DELETE);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path/id");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.body()).isNull();
+        assertThat(result.status()).isEqualTo(204);
+      }
+    }
+
+    @Nested
+    class PutTests {
+      @Test
+      public void shouldReturn200WithoutBody_whenEmptyPut(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(put("/path").willReturn(ok()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.PUT);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenPutBody(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(put("/path").willReturn(ok()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.PUT);
+        request.setHeaders(Map.of("header", "headerValue"));
+        request.setBody(Map.of("key1", "value1"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+
+        verify(
+            putRequestedFor(urlEqualTo("/path"))
+                .withHeader("Content-Type", equalTo("application/json"))
+                .withHeader("header", equalTo("headerValue"))
+
+   .withRequestBody(equalTo(StringEscapeUtils.unescapeJson("{\"key1\":\"value1\"}"))));
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenPutBodyURLEncoded(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(put("/path").willReturn(ok()));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.PUT);
+        request.setHeaders(
+            Map.of(HttpHeaders.CONTENT_TYPE,
+   ContentType.APPLICATION_FORM_URLENCODED.getMimeType()));
+        request.setBody(Map.of("key1", "value1", "key2", "value2"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+
+        verify(
+            putRequestedFor(urlEqualTo("/path"))
+                .withHeader("Content-Type", equalTo("application/x-www-form-urlencoded"))
+                .withRequestBody(
+                    and(containing("key1=value1"), containing("&"), containing("key2=value2"))));
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenPutBodyTextPlainWithStringBody(
+          WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        stubFor(put("/path").willReturn(ok().withBody("Hello, world updated!")));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.PUT);
+        request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE,
+   ContentType.TEXT_PLAIN.getMimeType()));
+        request.setBody("Hello, world!");
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        assertThat(result.body()).isEqualTo("Hello, world updated!");
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenPutBodyTextPlainWithIntegerBody(
+          WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        stubFor(put("/path").willReturn(ok().withBody("123")));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.PUT);
+        request.setHeaders(Map.of(HttpHeaders.CONTENT_TYPE,
+   ContentType.TEXT_PLAIN.getMimeType()));
+        request.setBody(123);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        assertThat(result.body()).isEqualTo("123");
+      }
+    }
+
+    @Nested
+    class AuthenticationTests {
+
+      @Test
+      public void shouldReturn200WithBody_whenGetWithBasicAuth(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(
+            get("/path")
+                .withBasicAuth("user", "password")
+                .willReturn(
+                    ok().withJsonBody(
+                            JsonNodeFactory.instance
+                                .objectNode()
+                                .put("name", "John")
+                                .put("age", 30)
+                                .putNull("message"))));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setHeaders(Map.of("Accept", "application/json"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        request.setAuthentication(new BasicAuthentication("user", "password"));
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        JSONAssert.assertEquals(
+            "{\"name\":\"John\",\"age\":30,\"message\":null}",
+            objectMapper.writeValueAsString(result.body()),
+            JSONCompareMode.STRICT);
+      }
+
+      @Test
+      public void shouldReturn401_whenGetWithWrongBasicAuth(WireMockRuntimeInfo wmRuntimeInfo) {
+        stubFor(
+            get("/path")
+                .withBasicAuth("user", "password")
+                .willReturn(unauthorized().withBody("Unauthorized")));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        request.setAuthentication(new BasicAuthentication("user", "password"));
+        ConnectorException e =
+            assertThrows(ConnectorException.class, () -> customApacheHttpClient.execute(request));
+        assertThat(e.getErrorCode()).isEqualTo("401");
+        assertThat(e.getMessage()).contains("Unauthorized");
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenGetWithBearerAuth(WireMockRuntimeInfo wmRuntimeInfo)
+          throws Exception {
+        stubFor(
+            get("/path")
+                .withHeader("Authorization", equalTo("Bearer token"))
+                .willReturn(
+                    ok().withJsonBody(
+                            JsonNodeFactory.instance
+                                .objectNode()
+                                .put("name", "John")
+                                .put("age", 30)
+                                .putNull("message"))));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setAuthentication(new BearerAuthentication("token"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        JSONAssert.assertEquals(
+            "{\"name\":\"John\",\"age\":30,\"message\":null}",
+            objectMapper.writeValueAsString(result.body()),
+            JSONCompareMode.STRICT);
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenGetWithApiKeyAuthInHeaders(
+          WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        stubFor(
+            get("/path")
+                .withHeader("theName", equalTo("theValue"))
+                .willReturn(
+                    ok().withJsonBody(
+                            JsonNodeFactory.instance
+                                .objectNode()
+                                .put("name", "John")
+                                .put("age", 30)
+                                .putNull("message"))));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setAuthentication(
+            new ApiKeyAuthentication(ApiKeyLocation.HEADERS, "theName", "theValue"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        JSONAssert.assertEquals(
+            "{\"name\":\"John\",\"age\":30,\"message\":null}",
+            objectMapper.writeValueAsString(result.body()),
+            JSONCompareMode.STRICT);
+      }
+
+      @Test
+      public void shouldReturn200WithBody_whenGetWithApiKeyAuthInQueryParams(
+          WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        stubFor(
+            get(urlPathEqualTo("/path"))
+                .withQueryParam("theName", equalTo("theValue"))
+                .willReturn(
+                    ok().withJsonBody(
+                            JsonNodeFactory.instance
+                                .objectNode()
+                                .put("name", "John")
+                                .put("age", 30)
+                                .putNull("message"))));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setAuthentication(
+            new ApiKeyAuthentication(ApiKeyLocation.QUERY, "theName", "theValue"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        JSONAssert.assertEquals(
+            "{\"name\":\"John\",\"age\":30,\"message\":null}",
+            objectMapper.writeValueAsString(result.body()),
+            JSONCompareMode.STRICT);
+      }
+
+      @ParameterizedTest
+      @ValueSource(strings = {OAuthConstants.BASIC_AUTH_HEADER, OAuthConstants.CREDENTIALS_BODY})
+      public void shouldReturn200WithBody_whenGetWithOAuthAndCredentialsInBody(
+          String credentialsLocation, WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        createAuthServer(credentialsLocation);
+        stubFor(
+            get("/path")
+                .withHeader("Authorization", equalTo("Bearer token"))
+                .willReturn(
+                    ok().withJsonBody(
+                            JsonNodeFactory.instance
+                                .objectNode()
+                                .put("name", "John")
+                                .put("age", 30)
+                                .putNull("message"))));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setAuthentication(
+            new OAuthAuthentication(
+                wmRuntimeInfo.getHttpBaseUrl() + "/oauth",
+                "clientId",
+                "clientSecret",
+                "theAudience",
+                credentialsLocation,
+                "read:resource"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        HttpCommonResult result = customApacheHttpClient.execute(request);
+        assertThat(result).isNotNull();
+        assertThat(result.status()).isEqualTo(200);
+        JSONAssert.assertEquals(
+            "{\"name\":\"John\",\"age\":30,\"message\":null}",
+            objectMapper.writeValueAsString(result.body()),
+            JSONCompareMode.STRICT);
+      }
+
+      @ParameterizedTest
+      @ValueSource(strings = {OAuthConstants.BASIC_AUTH_HEADER, OAuthConstants.CREDENTIALS_BODY})
+      public void shouldReturn401_whenGetWithOAuthReturns401(
+          String credentialsLocation, WireMockRuntimeInfo wmRuntimeInfo) throws Exception {
+        createFailingAuthServer(credentialsLocation);
+        stubFor(
+            get("/path")
+                .withHeader("Authorization", equalTo("Bearer token"))
+                .willReturn(
+                    ok().withJsonBody(
+                            JsonNodeFactory.instance
+                                .objectNode()
+                                .put("name", "John")
+                                .put("age", 30)
+                                .putNull("message"))));
+
+        HttpCommonRequest request = new HttpCommonRequest();
+        request.setMethod(HttpMethod.GET);
+        request.setAuthentication(
+            new OAuthAuthentication(
+                wmRuntimeInfo.getHttpBaseUrl() + "/oauth",
+                "clientId",
+                "clientSecret",
+                "theAudience",
+                credentialsLocation,
+                "read:resource"));
+        request.setUrl(wmRuntimeInfo.getHttpBaseUrl() + "/path");
+        var e = assertThrows(ConnectorException.class, () ->
+   customApacheHttpClient.execute(request));
+        assertThat(e).isNotNull();
+        assertThat(e.getErrorCode()).isEqualTo("401");
+        assertThat(e.getMessage()).contains("Unauthorized");
+      }
+
+      private void createAuthServer(String credentialsLocation) {
+        var request =
+            post("/oauth")
+                .withHeader(
+                    HttpHeaders.CONTENT_TYPE,
+                    equalTo(ContentType.APPLICATION_FORM_URLENCODED.getMimeType()))
+                .withFormParam("grant_type", equalTo("client_credentials"))
+                .withFormParam("audience", equalTo("theAudience"))
+                .withFormParam("scope", equalTo("read:resource"));
+        if (OAuthConstants.CREDENTIALS_BODY.equals(credentialsLocation)) {
+          request
+              .withFormParam("client_id", equalTo("clientId"))
+              .withFormParam("client_secret", equalTo("clientSecret"));
+        } else {
+          request.withBasicAuth("clientId", "clientSecret");
+        }
+        stubFor(
+            request.willReturn(
+                ok().withJsonBody(
+                        JsonNodeFactory.instance
+                            .objectNode()
+                            .put("access_token", "token")
+                            .put("token_type", "Bearer")
+                            .put("expires_in", 3600))));
+      }
+
+      private void createFailingAuthServer(String credentialsLocation) {
+        var request =
+            post("/oauth")
+                .withHeader(
+                    HttpHeaders.CONTENT_TYPE,
+                    equalTo(ContentType.APPLICATION_FORM_URLENCODED.getMimeType()))
+                .withFormParam("grant_type", equalTo("client_credentials"))
+                .withFormParam("audience", equalTo("theAudience"))
+                .withFormParam("scope", equalTo("read:resource"));
+        if (OAuthConstants.CREDENTIALS_BODY.equals(credentialsLocation)) {
+          request
+              .withFormParam("client_id", equalTo("clientId"))
+              .withFormParam("client_secret", equalTo("clientSecret"));
+        } else {
+          request.withBasicAuth("clientId", "clientSecret");
+        }
+        stubFor(request.willReturn(unauthorized().withBody("Unauthorized")));
+      }
+    }
 }
